+++ common/util/src/main/java/org/eclipse/dataspaceconnector/common/annotations/Uuid.java
@@ -1,0 +1,35 @@
+/*
+ *  Copyright (c) 2022 ZF Friedrichshafen AG
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       ZF Friedrichshafen AG - Initial Implementation
+ *
+ */
+
+package org.eclipse.dataspaceconnector.common.annotations;
+
+import jakarta.validation.Constraint;
+import jakarta.validation.Payload;
+import org.eclipse.dataspaceconnector.common.validator.UuidValidator;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Documented
+@Constraint(validatedBy = UuidValidator.class)
+@Target({ ElementType.METHOD, ElementType.FIELD })
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Uuid {
+    String message() default "Invalid UUID";
+    Class<?>[] groups() default {};
+    Class<? extends Payload>[] payload() default {};
+}